+++ packages/library-base/src/commonMain/kotlin/io/realm/dynamic/DynamicRealmObject.kt
@@ -16,8 +16,8 @@
 
 package io.realm.dynamic
 
+import io.realm.BaseRealmObject
 import io.realm.RealmList
-import io.realm.RealmObject
 import io.realm.schema.RealmStorageType
 import kotlin.reflect.KClass
 
@@ -26,7 +26,7 @@ import kotlin.reflect.KClass
  * based API instead of the conventional [Realm] API that only allows access through the properties
  * of the corresponding schema classes supplied in the configuration.
  */
-public interface DynamicRealmObject : RealmObject {
+public interface DynamicRealmObject : BaseRealmObject {
     /**
      * The type of the object.
      *