+++ packages/library-base/src/commonMain/kotlin/io/realm/internal/RealmObjectHelper.kt
@@ -16,8 +16,8 @@
 
 package io.realm.internal
 
+import io.realm.BaseRealmObject
 import io.realm.RealmList
-import io.realm.RealmObject
 import io.realm.dynamic.DynamicMutableRealmObject
 import io.realm.dynamic.DynamicRealmObject
 import io.realm.internal.interop.CollectionType
@@ -49,7 +49,7 @@ internal object RealmObjectHelper {
 
     @Suppress("unused") // Called from generated code
     internal inline fun getValue(
-        obj: RealmObjectReference<out RealmObject>,
+        obj: RealmObjectReference<out BaseRealmObject>,
         propertyName: String,
     ): RealmValue {
         obj.checkValid()
@@ -57,14 +57,14 @@ internal object RealmObjectHelper {
     }
 
     internal inline fun getValueByKey(
-        obj: RealmObjectReference<out RealmObject>,
+        obj: RealmObjectReference<out BaseRealmObject>,
         key: io.realm.internal.interop.PropertyKey,
     ): RealmValue = RealmInterop.realm_get_value(obj.objectPointer, key)
 
     // Return type should be R? but causes compilation errors for native
     @Suppress("unused") // Called from generated code
-    internal inline fun <reified R : RealmObject, U> getObject(
-        obj: RealmObjectReference<out RealmObject>,
+    internal inline fun <reified R : BaseRealmObject, U> getObject(
+        obj: RealmObjectReference<out BaseRealmObject>,
         propertyName: String,
     ): Any? {
         obj.checkValid()
@@ -74,12 +74,12 @@ internal object RealmObjectHelper {
     // Return type should be RealmList<R?> but causes compilation errors for native
     @Suppress("unused") // Called from generated code
     internal inline fun <reified R : Any> getList(
-        obj: RealmObjectReference<out RealmObject>,
+        obj: RealmObjectReference<out BaseRealmObject>,
         propertyName: String
     ): ManagedRealmList<Any?> = getList<R>(obj, propertyName, R::class)
 
     internal fun <R> getList(
-        obj: RealmObjectReference<out RealmObject>,
+        obj: RealmObjectReference<out BaseRealmObject>,
         propertyName: String,
         elementType: KClass<*>,
     ): ManagedRealmList<Any?> =
@@ -87,7 +87,7 @@ internal object RealmObjectHelper {
 
     // Cannot call managedRealmList directly from an inline function
     internal fun <R> getListByKey(
-        obj: RealmObjectReference<out RealmObject>,
+        obj: RealmObjectReference<out BaseRealmObject>,
         key: io.realm.internal.interop.PropertyKey,
         elementType: KClass<*>,
     ): ManagedRealmList<R> = getManagedRealmList(
@@ -119,7 +119,7 @@ internal object RealmObjectHelper {
     }
 
     @Suppress("unused") // Called from generated code
-    internal fun setValue(obj: RealmObjectReference<out RealmObject>, propertyName: String, value: RealmValue) {
+    internal fun setValue(obj: RealmObjectReference<out BaseRealmObject>, propertyName: String, value: RealmValue) {
         obj.checkValid()
         val key = obj.propertyInfoOrThrow(propertyName).key
         // TODO OPTIMIZE We are currently only doing this check for typed access so could consider
@@ -138,7 +138,7 @@ internal object RealmObjectHelper {
     }
 
     internal fun setValueByKey(
-        obj: RealmObjectReference<out RealmObject>,
+        obj: RealmObjectReference<out BaseRealmObject>,
         key: io.realm.internal.interop.PropertyKey,
         value: RealmValue,
     ) {
@@ -165,10 +165,10 @@ internal object RealmObjectHelper {
     }
 
     @Suppress("unused") // Called from generated code
-    internal inline fun <reified R : RealmObject> setObject(
-        obj: RealmObjectReference<out RealmObject>,
+    internal inline fun setObject(
+        obj: RealmObjectReference<out BaseRealmObject>,
         propertyName: String,
-        value: R?
+        value: BaseRealmObject?
     ) {
         obj.checkValid()
         val key = obj.propertyInfoOrThrow(propertyName).key
@@ -177,7 +177,7 @@ internal object RealmObjectHelper {
     }
 
     @Suppress("unused") // Called from generated code
-    internal inline fun <reified T : Any> setList(obj: RealmObjectReference<out RealmObject>, col: String, list: RealmList<Any?>) {
+    internal inline fun <reified T : Any> setList(obj: RealmObjectReference<out BaseRealmObject>, col: String, list: RealmList<Any?>) {
         val existingList = getList<T>(obj, col)
         if (list !is ManagedRealmList || !RealmInterop.realm_equals(existingList.nativePointer, list.nativePointer)) {
             existingList.also {
@@ -194,7 +194,7 @@ internal object RealmObjectHelper {
      * properties in the schema.
      */
     internal fun <R : Any> dynamicGet(
-        obj: RealmObjectReference<out RealmObject>,
+        obj: RealmObjectReference<out BaseRealmObject>,
         propertyName: String,
         clazz: KClass<R>,
         nullable: Boolean
@@ -207,7 +207,7 @@ internal object RealmObjectHelper {
         val value = when (clazz) {
             DynamicRealmObject::class,
             DynamicMutableRealmObject::class ->
-                realmValueToRealmObject(realmValue, clazz as KClass<out RealmObject>, obj.mediator, obj.owner)
+                realmValueToRealmObject(realmValue, clazz as KClass<out BaseRealmObject>, obj.mediator, obj.owner)
             else -> primitiveTypeConverters.getValue(clazz).realmValueToPublic(realmValue)
         }
         return value?.let {
@@ -221,7 +221,7 @@ internal object RealmObjectHelper {
     }
 
     internal fun <R : Any> dynamicGetList(
-        obj: RealmObjectReference<out RealmObject>,
+        obj: RealmObjectReference<out BaseRealmObject>,
         propertyName: String,
         clazz: KClass<R>,
         nullable: Boolean
@@ -232,12 +232,12 @@ internal object RealmObjectHelper {
         return getListByKey<R>(obj, propertyInfo.key, clazz) as RealmList<R?>
     }
 
-    internal fun <R> dynamicSetValue(obj: RealmObjectReference<out RealmObject>, propertyName: String, value: R) {
+    internal fun <R> dynamicSetValue(obj: RealmObjectReference<out BaseRealmObject>, propertyName: String, value: R) {
         obj.checkValid()
         val key = obj.propertyInfoOrThrow(propertyName).key
         val realmValue = when (value) {
             null -> RealmValue(null)
-            is RealmObject -> realmObjectToRealmValue(value, obj.mediator, obj.owner)
+            is BaseRealmObject -> realmObjectToRealmValue(value as BaseRealmObject, obj.mediator, obj.owner)
             else -> {
                 @Suppress("UNCHECKED_CAST")
                 (primitiveTypeConverters.getValue(value!!::class) as RealmValueConverter<Any>).publicToRealmValue(
@@ -248,11 +248,11 @@ internal object RealmObjectHelper {
         setValueByKey(obj, key, realmValue)
     }
 
-    private fun checkPropertyType(obj: RealmObjectReference<out RealmObject>, propertyName: String, collectionType: CollectionType, elementType: KClass<*>, nullable: Boolean): PropertyInfo {
+    private fun checkPropertyType(obj: RealmObjectReference<out BaseRealmObject>, propertyName: String, collectionType: CollectionType, elementType: KClass<*>, nullable: Boolean): PropertyInfo {
         val realElementType = when (elementType) {
             DynamicRealmObject::class,
             DynamicMutableRealmObject::class ->
-                RealmObject::class
+                BaseRealmObject::class
             else -> elementType
         }