+++ packages/library-base/src/commonMain/kotlin/io/realm/schema/RealmStorageType.kt
@@ -16,8 +16,8 @@
 
 package io.realm.schema
 
+import io.realm.BaseRealmObject
 import io.realm.RealmInstant
-import io.realm.RealmObject
 import kotlin.reflect.KClass
 
 /**
@@ -44,7 +44,7 @@ public enum class RealmStorageType(public val kClass: KClass<*>) {
     /**
      * Storage type for properties of type [RealmObject].
      */
-    OBJECT(RealmObject::class),
+    OBJECT(BaseRealmObject::class),
 
     /**
      * Storage type for properties of type [Float].