+++ CHANGELOG.md
@@ -1,7 +1,10 @@
 ## 0.12.0 (YYYY-MM-DD)
 
 ### Breaking Changes
-* None.
+* Introduced `BaseRealmObject` as base interface of `RealmObject` and `DynamicRealmObject` to prepare for future embedded object support.
+  * Most APIs accepts `BaseRealmObject` instead of `RealmObject`.
+  * `DynamicRealmObject` no longer implements `RealmObject` but only `BaseRealmObject`
+  * Besides the changes of base class of `DynamicRealmObject`, this should not require and code changes.
 
 ### Enhancements
 * None.